Milestone 5/requirements.txt
tensorflow==2.14.0
numpy==1.23.5
psycopg2-binary==2.9.9
Pillow==10.1.0
flask==3.0.1
